Created and adopted ElectronWaveImage class and corresponding docs and tests
